post

+++ app/Http/Controllers/TasksController.php
@@ -5,23 +5,36 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;
+use App\Http\Requests;
+use App\Http\Controllers\Controller;
 
 class TasksController extends Controller
 {
     public function index() {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks, 
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        } else {
+            return view('welcome');
+        }
     }
     
     public function show($id) {
-        $task = Task::find($id);
+        $user = \Auth::user();
+        $task = $user->tasks()->find($id);
         
-        return view('tasks.show', [
-            'task' => $task, 
-        ]);
+        $data = [
+            'user' => $user,
+            'task' => $task,
+        ];
+        return view('tasks.show', $data);
     }
     
     public function create() {
@@ -37,12 +50,19 @@ class TasksController extends Controller
             'status' => 'required', 
         ]);
         
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+        
+        /*
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        */
         
-        return redirect('/');
+        return redirect('/');   //==============================================
     }
     
     public function edit($id) {
@@ -54,22 +74,26 @@ class TasksController extends Controller
     }
     
     public function update(Request $request, $id) {
+
         $this->validate($request, [
             'status' => 'required', 
         ]);
         
-        $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->find($id)->update([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
         
-        return redirect('/');
+        return redirect('/');  //==============================================
     }
     
     public function destroy($id) {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+        
+        return redirect('/');  //==============================================
     }
 }
